+++ src/packages/property-editors/content-picker/components/input-tree/index.ts
