+++ src/packages/property-editors/content-picker/config/source-picker/manifests.ts
