+++ src/packages/property-editors/content-picker/config/source-picker/property-editor-ui-tree-picker-source-picker.element.ts
