+++ src/packages/property-editors/content-picker/config/source-type-picker/manifests.ts
