+++ src/packages/property-editors/content-picker/config/source-type-picker/property-editor-ui-tree-picker-source-type-picker.element.ts
