+++ src/packages/property-editors/content-picker/manifests.ts
