+++ src/packages/property-editors/content-picker/property-editor-ui-tree-picker.element.ts
